improved reliability with price of speed, and still not perfect
to fix : plot is changing painting offset to the end of file after was
scaled form big to very low zoom (exp. form 200% -> 10%)

+++ plot.h
@@ -75,7 +75,7 @@ private:
     //FFTFile
     int maxFFToffset;
     int offsetFileToOffsetFFT(quint32 offset) {return (offset*imgZoom) *((double)maxFFToffset/file->maxOffset());}
-    quint32 offsetFFTToOffsetFile(int offset) {return (offset/imgZoom) *(file->maxOffset()/maxFFToffset);}
+    quint32 offsetFFTToOffsetFile(int offset) {return (offset/imgZoom) *((double)file->maxOffset()/maxFFToffset);}
 
     //Painting
     virtual void paintEvent(QPaintEvent *);
